+++ packages/ragbits-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- feat: add get_vector_size method to all Embedders (#587)
+
 ## 0.19.1 (2025-05-27)
 
 ## 0.19.0 (2025-05-27)